refactor(react): reorganize modal files
