+++ packages/docsearch-react/src/DocSearchModal.tsx
@@ -3,25 +3,25 @@ import React, { type JSX } from 'react';
 
 import type { KeywordSearchBoxTranslations } from './components/KeywordSearchBox';
 import { KeywordSearchBox } from './components/KeywordSearchBox';
+import { ModalShell } from './components/ui/ModalShell';
 import type { DocSearchProps } from './DocSearch';
 import type { FooterTranslations } from './Footer';
 import { Footer } from './Footer';
 import { Hit } from './Hit';
-import { buildNoQuerySources, buildQuerySources, type BuildQuerySourcesState } from './modal/createDocSearchSources';
-import { DocSearchModalShell } from './modal/DocSearchModalShell';
-import { normalizeDocSearchIndexes } from './modal/normalizeDocSearchIndexes';
-import { useSendItemClickEvent } from './modal/useDocSearchInsights';
-import { useInitialModalQuery } from './modal/useInitialModalQuery';
-import { useModalEnvironment } from './modal/useModalEnvironment';
-import { useModalRefs } from './modal/useModalRefs';
-import { useRefreshOnInitialQuery } from './modal/useRefreshOnInitialQuery';
-import { useSaveRecentSearch } from './modal/useSaveRecentSearch';
-import { useStoredDocSearches } from './modal/useStoredDocSearches';
+import { useSendItemClickEvent } from './hooks/useDocSearchInsights';
+import { useInitialModalQuery } from './hooks/useInitialModalQuery';
+import { useModalEnvironment } from './hooks/useModalEnvironment';
+import { useModalRefs } from './hooks/useModalRefs';
+import { useRefreshOnInitialQuery } from './hooks/useRefreshOnInitialQuery';
+import { useSaveRecentSearch } from './hooks/useSaveRecentSearch';
+import { useStoredDocSearches } from './hooks/useStoredDocSearches';
 import type { ScreenStateTranslations } from './ScreenState';
 import { ScreenState } from './ScreenState';
 import type { DocSearchState, InternalDocSearchHit } from './types';
 import { useSearchClient } from './useSearchClient';
 import { identity, isModifierEvent, noop, scrollTo as scrollToUtils } from './utils';
+import { buildNoQuerySources, buildQuerySources, type BuildQuerySourcesState } from './utils/createDocSearchSources';
+import { normalizeDocSearchIndexes } from './utils/normalizeDocSearchIndexes';
 
 export type ModalTranslations = Partial<{
   searchBox: KeywordSearchBoxTranslations;
@@ -194,7 +194,7 @@ export function DocSearchModal({
   }
 
   return (
-    <DocSearchModalShell
+    <ModalShell
       state={state}
       containerRef={containerRef}
       modalRef={modalRef}
