+++ packages/docsearch-react/src/hooks/useDocSearchInsights.ts
