+++ packages/docsearch-react/src/hooks/useDocSearchModalEffects.ts
