+++ packages/docsearch-react/src/hooks/useInitialModalQuery.ts
