+++ packages/docsearch-react/src/hooks/useModalEnvironment.ts
