+++ packages/docsearch-react/src/hooks/useModalRefs.ts
