+++ packages/docsearch-react/src/hooks/useRefreshOnInitialQuery.ts
