+++ packages/docsearch-react/src/hooks/useSaveRecentSearch.ts
