+++ packages/docsearch-react/src/hooks/useStoredDocSearches.ts
