+++ packages/docsearch-react/src/utils/createAskAiSources.ts
