+++ packages/docsearch-react/src/utils/normalizeDocSearchIndexes.ts
